fixed bugs: cannot shoot in map bounds we can rotate and shoot when we in pause state guns spawning more and more after  Pause->Menu->Play many coroutine attack player from one enemy shotgun shooting once more after changing weapon (when reload coroutine is continue)

+++ Assets/scripts/core/activeObjects/player/Player.cs
@@ -9,7 +9,7 @@ namespace Global.Player
 #pragma warning disable
 
         [SerializeField] private int hpMaximum;
-        [SerializeField] private int speed;
+        [SerializeField] private float speed;
         [SerializeField] private int defence;
         [SerializeField] private float hpValue;
 
@@ -38,7 +38,7 @@ namespace Global.Player
             hpValue -= hp;
         }
 
-        public void SetPlayerStatFromData(int hp, int speed, int defence)
+        public void SetPlayerStatFromData(int hp, float speed, int defence)
         {
             hpMaximum = hp;
             this.defence = defence;
